[LCR-14] rename: auth 도메인 변경
