+++ src/main/java/com/lgcns/domain/auth/dto/Request/ManagerCreateRequest.java
@@ -1,4 +1,4 @@
-package com.lgcns.domain.manager.dto.Request;
+package com.lgcns.domain.auth.dto.Request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
