+++ src/main/java/com/lgcns/domain/auth/service/ManagerService.java
@@ -1,0 +1,7 @@
+package com.lgcns.domain.auth.service;
+
+import com.lgcns.domain.auth.dto.Request.ManagerCreateRequest;
+
+public interface ManagerService {
+    void createManager(ManagerCreateRequest request);
+}
