+++ src/main/java/com/lgcns/domain/auth/service/ManagerServiceImpl.java
@@ -1,9 +1,9 @@
-package com.lgcns.domain.manager.service;
+package com.lgcns.domain.auth.service;
 
-import com.lgcns.domain.manager.domain.Manager;
-import com.lgcns.domain.manager.dto.Request.ManagerCreateRequest;
-import com.lgcns.domain.manager.exception.ManagerErrorCode;
-import com.lgcns.domain.manager.repository.ManagerRepository;
+import com.lgcns.domain.auth.domain.Manager;
+import com.lgcns.domain.auth.dto.Request.ManagerCreateRequest;
+import com.lgcns.domain.auth.exception.ManagerErrorCode;
+import com.lgcns.domain.auth.repository.ManagerRepository;
 import com.lgcns.global.error.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
