--- src/main/java/com/lgcns/domain/manager/service/ManagerService.java
@@ -1,7 +1,0 @@
-package com.lgcns.domain.manager.service;
-
-import com.lgcns.domain.manager.dto.Request.ManagerCreateRequest;
-
-public interface ManagerService {
-    void createManager(ManagerCreateRequest request);
-}
